requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.10.0
nltk>=3.6.0
matplotlib>=3.4.0
seaborn>=0.11.0
argparse>=1.4.0
regex>=2021.8.3
tqdm>=4.62.0
requests>=2.28.0
python-dotenv>=0.19.0 